update pay method.amount 0.10 to 0.1 for sign

+++ ehcore/src/main/java/com/everhomes/organization/pm/PropertyMgrServiceImpl.java
@@ -3627,7 +3627,7 @@ public class PropertyMgrServiceImpl implements PropertyMgrService {
 		map.put("appKey",appKey);
 		map.put("timestamp",timestamp+"");
 		map.put("randomNum",randomNum+"");
-		map.put("amount",dto.getAmount()+"");
+		map.put("amount",dto.getAmount().doubleValue()+"");
 		String signature = SignatureHelper.computeSignature(map, app.getSecretKey());
 		dto.setAppKey(appKey);
 		dto.setRandomNum(randomNum);
