+++ ehcore/src/main/java/com/everhomes/techpark/park/ParkServiceImpl.java
@@ -232,7 +232,7 @@ public class ParkServiceImpl implements ParkService {
 		map.put("appKey",appKey);
 		map.put("timestamp",timestamp+"");
 		map.put("randomNum",randomNum+"");
-		map.put("amount",dto.getAmount()+"");
+		map.put("amount",dto.getAmount().doubleValue()+"");
 		String signature = SignatureHelper.computeSignature(map, app.getSecretKey());
 		dto.setAppKey(appKey);
 		dto.setRandomNum(randomNum);
